+++ qupath-core-processing/src/main/java/qupath/opencv/dnn/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * QuPath classes to use OpenCV's DNN module.
+ */
+package qupath.opencv.dnn;